Move upload image components from src to components
