+++ components/atoms/error-message/index.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import ErrorMessage from '_atoms/error-message'
+import ErrorMessage from '_components/atoms/error-message'
 
 storiesOf('Atoms/ErrorMessage', module).add('Error Message', () => (
   <ErrorMessage>This is an error message</ErrorMessage>
