+++ components/atoms/loader/index.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import Loader from '_atoms/loader'
+import Loader from '_components/atoms/loader'
 
 storiesOf('Atoms/Loader', module).add('Loader', () => <Loader />)
