+++ components/atoms/overlay/index.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import Overlay from '_atoms/overlay'
+import Overlay from '_components/atoms/overlay'
 
 storiesOf('Atoms/Overlay', module).add('Overlay', () => <Overlay show />)
