+++ components/molecules/avatar/index.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import Avatar from '_molecules/avatar'
+import Avatar from '_components/molecules/avatar'
 
 storiesOf('Molecules/Avatar', module).add('Avatar', () => <Avatar />)
