+++ components/organisms/dialog-capture-image/styles.css
@@ -1,0 +1,6 @@
+.uploadButton {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  margin-top: 15px;
+}
