+++ components/organisms/dialog-preview-image/styles.css
