+++ components/organisms/editable-image/styles.css
