--- src/components/atoms/button/styles.css
@@ -1,8 +1,0 @@
-.button {
-  background-color: var(--color-primary);
-  color: var(--color-white);
-  width: 174px;
-  border: none;
-  padding: 6px;
-  margin-top: 10px;
-}
