--- src/components/molecules/icon-button/styles.css
@@ -1,15 +1,0 @@
-.button {
-  display: inline-flex;
-  height: 40px;
-  width: 40px;
-  background: transparent;
-  border: 0;
-  cursor: pointer;
-  padding: 0;
-  align-items: center;
-  justify-content: center;
-
-  &:focus {
-    outline: 0;
-  }
-}
